Widen InventoryUserUIInterfaceEntry stack indices to 16 bits
Each entry in an inventory menu lists the stacks it shows, by each stack's position in the item's stack list. One entry can cover several stacks. The engine stores each position in 2 bytes, but they were declared as 1 byte. So stackIndex[1] was read just 1 byte into the list, which is still inside the first position, and usually came out as 0. A position above 255 needs both of its bytes, so it came out wrong too.

The game's own code reads these numbers 2 bytes at a time, both in InventoryItemDisplayData's constructor and in the function that adds up an entry's item count. In the disassembly this shows up as word ptr reads, and a word is 2 bytes.

+++ include/RE/I/InventoryUserUIInterfaceEntry.h
@@ -9,8 +9,8 @@ namespace RE
 	{
 	public:
 		// members
-		InventoryInterface::Handle     invHandle;   // 00
-		BSTSmallArray<std::uint8_t, 4> stackIndex;  // 08
+		InventoryInterface::Handle      invHandle;   // 00
+		BSTSmallArray<std::uint16_t, 4> stackIndex;  // 08
 	};
 	static_assert(sizeof(InventoryUserUIInterfaceEntry) == 0x20);
 }
